requirements.txt
streamlit==1.39.0
langchain==0.3.7
langchain-community==0.3.7
langchain-ollama==0.2.0
chromadb==0.5.5
pypdf==4.3.1
tiktoken==0.8.0
python-dotenv==1.0.1